requirements.txt
pandas
PyYAML
matplotlib
seaborn
numpy
tensorflow
scikit-learn